+++ sdl3-sys-gen/src/emit/patch.rs
@@ -242,6 +242,28 @@ pub fn patch_emit_define(ctx: &mut EmitContext, define: &Define) -> Result<bool,
 
 type EmitDefinePatch = EmitPatch<Define>;
 
+#[allow(unused_macros)]
+macro_rules! emit_define_alias {
+    ($module:literal, $new:literal, $old:literal, $version:literal) => {
+        EmitDefinePatch {
+            module: Some($module),
+            match_ident: |i| i == $new,
+            patch: |ctx, d| {
+                d.emit(ctx)?;
+                writeln!(
+                    ctx,
+                    r#"#[deprecated(since = "{}", note = "renamed to `{}`")]"#,
+                    $version, $new,
+                )?;
+                let mut d = d.clone();
+                d.ident = Ident::new_inline($old).into();
+                d.emit(ctx)?;
+                Ok(true)
+            },
+        }
+    };
+}
+
 const EMIT_DEFINE_PATCHES: &[EmitDefinePatch] = &[
     EmitDefinePatch {
         // skip emitting these