+++ sdl3-ttf-sys/ChangeLog.md
@@ -3,6 +3,7 @@
 - 0.3.0 WIP:
     - Update sdl3-sys to 0.6.0
     - Add metadata
+    - Remove deprecated aliases
 
 - 0.2.0:
     - Update sdl3-sys to 0.5.0